#169 micromamba conda soft switch
Implemented a version of album which can use both conda and micromamba. Micromamba is the preferred package manager, but can only be utilized when micromamba was installed in the .album directory. 
For the use of micromamba there is now a new controller, the MicromambaManager which inherits the functionality of the conda manager but alters a few function to fit the syntax of micromamba.
Also changed the name of the conda_maganer in the EnvironmentManager class to package_manager to avoid confusion when using micromamba.