Fix for old collections
With the new migration manager the location of the collection.json file changed. If it was not moved to the new place album threw an error instead of moving the json file. Not if it is missing it will be checked if it is still ln the old location and moved to the correct one before throwing an error.