Fixed an LogLevel Error
The Loglevel switch for IntEnum to Enum caused a problem in if statements. You can use an IntEnum object in a condition without acessing its .value . This you cannot do with a normal Enum object. 

For example:
if IntEnum: is equvalent to if IntEnum.value

This is not the case for 
if Enum      or     if  Enum.value