Fixed test solution
Due to a branching error of mine in the album runner repo, the runner git the feature of creating boolean values in the run scripts without ''. This caused the value of DefaultbooleanValue.txt to be handled as a boolean value instead of the string it really was. Now its a boolean value and there will be no more problems with ne the new runner feature.